@@ -1,12 +1,12 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Bot.Builder;
 using Microsoft.Bot.Builder.Integration.AspNet.Core;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Core.Configuration;
 using NewHost.Bots;
+using Microsoft.Extensions.Hosting;
 
 namespace NewHost
 {
@@ -21,27 +21,32 @@ namespace NewHost
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddControllers();
+            services.AddHealthChecks();
+            services.AddApplicationInsightsTelemetry();
             services.AddSingleton<IBotFrameworkHttpAdapter, AdapterWithErrorHandler>();
             services.AddTransient<IBot, HorssonBot>();
             services.RegisterCoreDependencies();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-            }
-            else
+            } else
             {
+                app.UseHttpsRedirection();
                 app.UseHsts();
             }
-
-            app.UseDefaultFiles();
-            app.UseStaticFiles();
+            
+            app.UseRouting();
             app.UseWebSockets();
-            app.UseMvc();
+            app.UseHealthChecks("/health");
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
